+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Maps2: Updated Amenities selection support [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 ### Added
 - Maps2: added analytics events [#5344](https://github.com/rokwire/illinois-app/issues/5344).
+- Maps2: added select location panel [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 ### Fixed
 - Handle tap action over the image in GroupCard for "All" and "Pending" groups.